packages/api/pyproject.toml
[project]
name = "dialectica-api"
version = "1.0.0"
description = "DIALECTICA FastAPI Service — The Universal Data Layer for Human Friction"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [{ name = "TACITUS", email = "team@tacitus.ai" }]
dependencies = [
    "dialectica-ontology",
    "dialectica-graph",
    "dialectica-extraction",
    "dialectica-reasoning",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic-settings>=2.5",
    "python-multipart>=0.0.12",
    "google-cloud-pubsub>=2.26",
    "google-cloud-storage>=2.18",
    "google-cloud-secret-manager>=2.20",
    "prometheus-fastapi-instrumentator>=7.0",
    "prometheus-client>=0.21",
    "redis[hiredis]>=5.2",
    "structlog>=24.4",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-gcp-trace>=1.8",
    "pyyaml>=6.0",
    "sqlmodel>=0.0.18",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "python-ulid>=2.0",
    "bcrypt>=4.1",
    "graphiti-core>=0.29.0",
    "cozo-embedded>=0.7.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dialectica_api"]
